Remove Gst dependency from camera.py and to use video_pipeline.py
