+++ surface/video/camera.py
@@ -2,42 +2,30 @@ import cv2
 import numpy as np
 from PIL import Image
 from io import BytesIO
-import gi
-gi.require_version('Gst', '1.0')
-gi.require_version('GstApp', '1.0')
-from gi.repository import Gst, GstApp
 
 import get_image
+import video_pipeline
 
 FRAMES_TO_SAVE = 2  # save a frame every x frames
 
 class Camera:
     def __init__(self, port):
         self._frame = None
-        Gst.init(None)
         gst_str = f'udpsrc port={port} ! application/x-rtp ! rtpjitterbuffer ! rtph264depay ! avdec_h264 ! videoconvert ! video/x-raw,width=1280,height=720,format=RGB ! appsink name=appsink0'
         print(gst_str)
-        self.pipeline = Gst.parse_launch(gst_str)
-        self.appsink = self.pipeline.get_by_name('appsink0')
-        self.appsink.set_property('emit-signals', True)
-        self.appsink.set_property('max-buffers', 1)
-        self.appsink.set_property('drop', True)
-        self.appsink.set_property('sync', False)
-        self.appsink.connect('new-sample', self.on_new_buffer)
-
+        video_pipeline.setup(gst_str)
+        video_pipeline.connect_sink_listener(self.on_new_buffer)
+        
         self.is_saving = False
         self.frame_counter = 0
         self.last_saved_frame = 0
 
-        self.pipeline.set_state(Gst.State.PLAYING)
-
     def __del__(self):
-        self.pipeline.set_state(Gst.State.NULL)
+        video_pipeline.teardown()
 
     def start_saving(self):
         self.is_saving = True
         # get_image.make_folder()
-    
     def stop_saving(self):
         self.is_saving = False
 
@@ -50,7 +38,6 @@ class Camera:
             dtype=np.uint8,
         )
         self.frame_counter += 1
-        return Gst.FlowReturn.OK
 
     def get_frame(self) -> bytes: ## makes image from array of values and stores as jpeg
         if self._frame is not None and self._frame.any():
